# api/swagger.yaml
basePath: /api/v1
definitions:
  domain.Address:
    properties:
      city:
        type: string
      country:
        type: string
      id:
        type: string
      street:
        type: string
    type: object
  domain.Error:
    properties:
      key:
        type: string
      value:
        type: string
    type: object
  domain.Image:
    properties:
      data:
        items:
          type: integer
        type: array
      hash:
        type: string
      id:
        type: string
    type: object
  domain.Product:
    properties:
      available_stock:
        type: integer
      category:
        type: string
      id:
        type: string
      image:
        $ref: '#/definitions/domain.Image'
      last_update_date:
        type: string
      name:
        type: string
      price:
        type: number
      supplier:
        $ref: '#/definitions/domain.Supplier'
    type: object
  domain.Supplier:
    properties:
      address:
        $ref: '#/definitions/domain.Address'
      id:
        type: string
      name:
        type: string
      phone_number:
        type: string
    type: object
...
  dto.Product:
    properties:
      available_stock:
        type: integer
      category:
        type: string
      city:
        type: string
      country:
        type: string
      image:
        items:
          type: integer
        type: array
      name:
        type: string
      phone_number:
        type: string
      price:
        type: number
      street:
        type: string
    required:
    - available_stock
    - category
    - image
    - name
    - name
    - phone_number
    - price
    type: object
...
host: aboba.com
info:
  contact: {}
  description: This is a sample server Petstore server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Swagger CRUD Home appliance store API
  version: "1.0"
paths:
...
  /api/v1/products:
    get:
      consumes:
      - application/json
      description: The endpoint for retrieve all registered product in system
      parameters:
      - description: limit get product
        in: query
        name: limit
        type: integer
      - description: offset get product
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.Product'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Error'
      summary: Get all product
      tags:
      - products
    post:
      consumes:
      - application/json
      description: 'Product created from JSON or XML, for create endpoint required:
        name, category, price, available_stock, supplier_name, supplier_phone_number,
        image'
      parameters:
      - description: Product data
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/domain.Product'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Error'
      summary: Create product
      tags:
      - products
  /api/v1/products/{id}:
    delete:
      consumes:
      - application/json
      description: The endpoint for deleting product data by ID
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Error'
      summary: Delete product by ID
      tags:
      - products
    get:
      consumes:
      - application/json
      description: The endpoint for retrieve registered product in system by id
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Error'
      summary: Get product by id
      tags:
      - products
    patch:
      consumes:
      - application/json
      description: The endpoint for updating product data (avaliable stock) by ID
        to a decrease avalible stock
      parameters:
      - description: Decrease value
        in: query
        name: decrease
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/domain.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/domain.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/domain.Error'
      summary: Update product by ID
      tags:
      - products
...
swagger: "2.0"